Re-writing snare with latest web server
